Allow 16bpp gfxdecode (#5167)
* digfx.cpp : Add 16bpp case of RAW gfx layout
drawgfx.cpp, tilemap.cpp : Make changeable total elements of gfx_elements constructor at RAW case, Allow 16bpp gfxdecode, Fix spacing
tilemap.cpp : Allow 16bpp tilemap pen data
gstream.cpp : Use gfxdecode for 16bpp gfx of X2222, Fix spacing
igs017_igs031.cpp : Cleanup single sprite drawing routine
jedi.cpp : Improve debug gfxdecode viewer(text gfx)

* cps1.cpp : Minor fixes

* gstream.cpp : Fix regression in x2222
drawgfx.cpp : Fix 16bpp transparent pen

* gstream.cpp : Fix regression, Reduce unnecessary lines

* Sync to master

+++ src/emu/drawgfxm.h
@@ -470,7 +470,7 @@ while (0)
 do {                                                                                \
 	g_profiler.start(PROFILER_DRAWGFX);                                             \
 	do {                                                                            \
-		const u8 *srcdata;                                                          \
+		const u16 *srcdata;                                                         \
 		s32 destendx, destendy;                                                     \
 		s32 srcx, srcy;                                                             \
 		s32 curx, cury;                                                             \
@@ -549,7 +549,7 @@ do {                                                                            
 			{                                                                       \
 				PRIORITY_TYPE *priptr = PRIORITY_ADDR(priority, PRIORITY_TYPE, cury, destx); \
 				PIXEL_TYPE *destptr = &dest.pixt<PIXEL_TYPE>(cury, destx);          \
-				const u8 *srcptr = srcdata;                                         \
+				const u16 *srcptr = srcdata;                                        \
 				srcdata += dy;                                                      \
 																					\
 				/* iterate over unrolled blocks of 4 */                             \
@@ -584,7 +584,7 @@ do {                                                                            
 			{                                                                       \
 				PRIORITY_TYPE *priptr = PRIORITY_ADDR(priority, PRIORITY_TYPE, cury, destx); \
 				PIXEL_TYPE *destptr = &dest.pixt<PIXEL_TYPE>(cury, destx);          \
-				const u8 *srcptr = srcdata;                                         \
+				const u16 *srcptr = srcdata;                                        \
 				srcdata += dy;                                                      \
 																					\
 				/* iterate over unrolled blocks of 4 */                             \
@@ -642,7 +642,7 @@ do {                                                                            
 do {                                                                                \
 	g_profiler.start(PROFILER_DRAWGFX);                                             \
 	do {                                                                            \
-		const u8 *srcdata;                                                          \
+		const u16 *srcdata;                                                         \
 		u32 dstwidth, dstheight;                                                    \
 		s32 destendx, destendy;                                                     \
 		s32 srcx, srcy;                                                             \
@@ -730,7 +730,7 @@ do {                                                                            
 		{                                                                           \
 			PRIORITY_TYPE *priptr = PRIORITY_ADDR(priority, PRIORITY_TYPE, cury, destx); \
 			PIXEL_TYPE *destptr = &dest.pixt<PIXEL_TYPE>(cury, destx);              \
-			const u8 *srcptr = srcdata + (srcy >> 16) * rowbytes();                 \
+			const u16 *srcptr = srcdata + (srcy >> 16) * rowbytes();                \
 			s32 cursrcx = srcx;                                                     \
 			srcy += dy;                                                             \
 																					\
