+++ src/mame/drivers/mediagx.cpp
@@ -276,7 +276,7 @@ void mediagx_state::video_start()
 void mediagx_state::draw_char(bitmap_rgb32 &bitmap, const rectangle &cliprect, gfx_element *gfx, int ch, int att, int x, int y)
 {
 	int i,j;
-	const uint8_t *dp;
+	const uint16_t *dp;
 	int index = 0;
 	const pen_t *pens = &m_palette->pen(0);
 
@@ -287,7 +287,7 @@ void mediagx_state::draw_char(bitmap_rgb32 &bitmap, const rectangle &cliprect, g
 		uint32_t *p = &bitmap.pix32(j);
 		for (i=x; i < x+8; i++)
 		{
-			uint8_t pen = dp[index++];
+			uint16_t pen = dp[index++];
 			if (pen)
 				p[i] = pens[gfx->colorbase() + (att & 0xf)];
 			else
